feat: allow orientation manager to edit any notebook from its deployment
