+++ app/src/routes/(auth)/orientation/carnets/[notebook_id]/+page.svelte
@@ -1,29 +1,31 @@
 <script lang="ts">
-	import { GetNotebookByIdDocument } from '$lib/graphql/_gen/typed-document-nodes';
-	import { displayFullName } from '$lib/ui/format';
+	import { GetNotebookDocument } from '$lib/graphql/_gen/typed-document-nodes';
+	import NotebookEdit from '$lib/ui/OrientationManager/NotebookEdit.svelte';
 	import LoaderIndicator from '$lib/ui/utils/LoaderIndicator.svelte';
-	import NotebookView from '$lib/ui/views/NotebookView.svelte';
 	import { operationStore, query } from '@urql/svelte';
 	import type { PageData } from './$types';
 
 	export let data: PageData;
 
-	const getNotebookResult = operationStore(GetNotebookByIdDocument, {
-		id: data.notebookId,
-	});
-	query(getNotebookResult);
+	const getNotebookResult = operationStore(
+		GetNotebookDocument,
+		{ id: data.notebookId },
+		{
+			additionalTypenames: [
+				'beneficiary',
+				'notebook_member',
+				'wanted_job',
+				'notebook_appointment',
+				'notebook_focus',
+				'notebook_action',
+			],
+		}
+	);
 
-	$: name = $getNotebookResult.data?.notebook
-		? displayFullName($getNotebookResult.data?.notebook.beneficiary)
-		: '';
+	query(getNotebookResult);
+	$: notebook = $getNotebookResult.data?.notebook;
 </script>
 
-<svelte:head>
-	<title>Carnet {name} - Carnet de bord</title>
-</svelte:head>
-
 <LoaderIndicator result={getNotebookResult}>
-	<div class="my-6">
-		<NotebookView notebook={$getNotebookResult.data.notebook} />
-	</div>
+	<NotebookEdit {notebook} />
 </LoaderIndicator>
