--- app/src/routes/(auth)/orientation/carnets/edition/[notebook_id]/+page.ts
@@ -1,8 +1,0 @@
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = ({ params }) => {
-	const notebookId = params.notebook_id;
-	return {
-		notebookId,
-	};
-};
