+++ hasura/metadata/databases/carnet_de_bord/tables/public_notebook.yaml
@@ -386,12 +386,9 @@ update_permissions:
         - work_situation_date
         - work_situation_end_date
       filter:
-        members:
-          _and:
-            - account_id:
-                _eq: X-Hasura-User-Id
-            - active:
-                _eq: true
+        beneficiary:
+          deployment_id:
+            _eq: X-Hasura-Deployment-Id
       check: null
   - role: professional
     permission:
